Moved page object code into separate files
