+++ e2e-tests/location.js
@@ -1,31 +1,28 @@
 /**
  * End-to-end tests for location-related content.
  */
-const webdriver = require('selenium-webdriver');
 const test = require('selenium-webdriver/testing');
 const chai = require('chai');
 const expect = chai.expect;
 const chaiAsPromised = require('chai-as-promised');
 chai.use(chaiAsPromised);
 const config = require('./support/e2e-tests.config.js');
-
+const LocationLookupPage = require('./location-lookup-page');
 const url = config.getBaseUrl();
-const By = webdriver.By;
-const until = webdriver.until;
 
 test.describe('location lookup', function() {
   let page = undefined;
   test.beforeEach(function() {
     this.driver.get(url);
-    page = new LookupLocationPage(this.driver);
+    page = new LocationLookupPage(this.driver);
   });
 
   test.afterEach(function() {
     page = undefined;
   });
 
-  test.it('Should lookup location', function() {
-    const zip = '12222';
+  test.it('Should lookup location by zip code', function() {
+    const location = '12222';
 
     // give the IP lookup a chance to complete;
     //  otherwise it interferes with the location lookup
@@ -34,51 +31,50 @@ test.describe('location lookup', function() {
     // click on location to display location input box
     page.displayLocationInputBox();
 
-    // enter zip in address text box and submit
-    page.enterNewLocation(zip);
+    // enter location in address text box and submit
+    page.enterAndSubmitNewLocation(location);
 
-    //check for zip code link
-    const newButtonText = page.getNewLocationButtonText(zip);
-    return expect(newButtonText).to.eventually.equal(zip);
+    // check for location text
+    const newLocationText = page.getNewLocationElement(location);
+    return expect(newLocationText.getText()).to.eventually.equal(location);
+  });
+
+  test.it('Should lookup location by location name', function() {
+    const location = 'GARY, IN';
+
+    // give the IP lookup a chance to complete;
+    //  otherwise it interferes with the location lookup
+    this.driver.sleep(2000);
+
+    // click on location to display location input box
+    page.displayLocationInputBox();
+
+    // enter location in address text box and submit
+    page.enterAndSubmitNewLocation(location);
+
+    // check for location text
+    const newLocationText = page.getNewLocationElement(location);
+    return expect(newLocationText.getText()).to.eventually.equal(location);
+  });
+
+  test.it('Should show error message if lookup location is invalid', function() {
+    const location = 'asdf'; // bogus location
+    const expected = page.getLocationErrorMessage();
+
+    // give the IP lookup a chance to complete;
+    //  otherwise it interferes with the location lookup
+    this.driver.sleep(2000);
+
+    // click on location to display location input box
+    page.displayLocationInputBox();
+
+    // enter location in address text box and submit
+    page.enterAndSubmitNewLocation(location);
+
+    // check for expected error message
+    const newLocationText = page.getNewLocationElement(expected);
+    return expect(newLocationText.getText()).to.eventually.equal(expected);
   });
 
 });
 
-/**
- * Page object for location-related content.
- *
- * @class LookupLocationPage
- */
-class LookupLocationPage {
-  constructor(driver) {
-    this.driver = driver;
-    // define selectors
-    this.locationButtonSelector = 'button.subtle-button';
-    this.locationInputSelector = '#address';
-    this.locationSubmitSelector = 'p > form > button';
-  }
-
-  displayLocationInputBox() {
-    const selector = By.css(this.locationButtonSelector);
-    this.driver.wait(until.elementLocated(selector), 5000);
-    const addressButton = this.driver.findElement(selector);
-    addressButton.click();
-  }
-
-  enterNewLocation(location) {
-    const inputSelector = By.css(this.locationInputSelector);
-    const submitSelector = By.css(this.locationSubmitSelector);
-    this.driver.wait(until.elementLocated(inputSelector), 10000);
-    this.driver.findElement(inputSelector).sendKeys(location);
-    // this.driver.findElement(submitSelector).sendKeys(Key.ENTER);
-    this.driver.findElement(submitSelector).click();
-  }
-
-  getNewLocationButtonText(location) {
-    const selector = By.css(this.locationButtonSelector);
-    this.driver.wait(until.elementLocated(selector), 6000);
-    const addressButton = this.driver.findElement(selector);
-    this.driver.wait(until.elementTextIs(addressButton, location), 2000);
-    return addressButton.getText();
-  }
-}
